Add a lower level conv interface that doesn't use keyword arguments
The keyword arguments to the various conv arguments affect the output
shape. If the arrays track their shape, this puts significant strain
on the compiler to track the keyword arguments through the various
layers of canonicalization (as well as needing to be captured by
value in Zygote's derivative code). To make things easier on the
compiler, create a couple of structs that behave as functions, but
have type parameters to configure the convolution. The compiler
is much happier about that and can successfully infer Zygote's
gradient function (after both Flux and Zygote are changed to
use this interface instead of the keyword arguments of course).
I don't think it's particularly pretty, but it seems to work
fine for now and we can fix it as part of a larger overhaul
of NNlib's intrinsics. To avoid disrupting users, I keep the
keyword argument versions as well (though to get the type
inference benefits one needs to avoid using those).