Rewrite `im2col()` for greater performance
The performance gain of this PR comes mainly from identifying that the
conditionals within the `im2col()` inner loops are really only needed
towards the edges of the image being `im2col()`'ed.  By separating the
loops into two separate segments, we are able to achieve very high
memory bandwidth utilizaiton for the majority of the wallclock time of
the function.  Eventually we will want to parallelize this loop via
threading, but until that is possible, this should be a nice
improvement. In my testing, this speeds up `im2col()` by a factor of
2-3x.

This change also begins the work of pushing some convolutional
parameters into the type domain.  This is especially helpful for
`im2col()`, as it makes it possible that the compiler could completely
elide the second set of for loops in the case of no padding (this is not
tested; I don't know if the compiler actually is able to elide the loops)
but pushing these parameters into the type domain gives a substantial
speed boost.