Big interface overhaul
Eliminate as much runtime checking as possible; create metaprogramming
forwards from frontend to backend in a better (yet still imperfect) manner.

We are reserving the right for the frontend names such as `conv()` to
decide which backend to use at runtime, but for now, all `conv()` calls
(but not all pooling calls) are dispatched at compile-time, which is nice.